customized

+++ ws2811.go
@@ -33,7 +33,7 @@ See https://github.com/jgarff/rpi_ws281x for instructions
 package ws2811
 
 /*
-#cgo CFLAGS: -std=c99
+#cgo CFLAGS: -std=c99 -D_XOPEN_SOURCE=600 -Wno-implicit-function-declaration
 #cgo LDFLAGS: -lws2811
 #include "ws2811.go.h"
 */
@@ -47,7 +47,7 @@ import (
 func Init(gpioPin int, ledCount int, brightness int) error {
 	C.ledstring.channel[0].gpionum = C.int(gpioPin)
 	C.ledstring.channel[0].count = C.int(ledCount)
-	C.ledstring.channel[0].brightness = C.uint8_t(brightness)
+	C.ledstring.channel[0].brightness = C.int(brightness)
 	res := int(C.ws2811_init(&C.ledstring))
 	if res == 0 {
 		return nil
